requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
black==22.12.0
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.3
mypy-extensions==0.4.3
passlib==1.7.4
password-validator==1.0
pathspec==0.10.3
platformdirs==2.6.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
yarl==1.8.2
